Tidy up

+++ src/Console/Commands/Setup.php
@@ -4,18 +4,16 @@
 
 namespace SebKay\Touchstone\Console\Commands;
 
-use SebKay\Touchstone\SQLiteConnection;
+use const SebKay\Touchstone\SQLITE_FILE_PATH;
 
+use SebKay\Touchstone\SQLiteConnection;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 use function Laravel\Prompts\password;
-
 use function Laravel\Prompts\text;
-
-use const SebKay\Touchstone\SQLITE_FILE_PATH;
 
 #[AsCommand(name: 'setup')]
 class Setup extends Command
@@ -51,23 +49,31 @@ class Setup extends Command
 
         // \ray($db);
 
-        \unlink(SQLITE_FILE_PATH);
-        \touch(SQLITE_FILE_PATH);
+        $this->resetDatabase();
 
         try {
             $pdo = (new SQLiteConnection())->connect();
             // create table
-            $tableName = 'touchstone_' . \random_int(1000, 9999);
+            $tableName = 'touchstone_'.\random_int(1000, 9999);
             $pdo->exec("CREATE TABLE $tableName (id INTEGER PRIMARY KEY, name TEXT)");
 
             // get list of tables
             $tables = $pdo->query("SELECT name FROM sqlite_master WHERE type='table' ORDER BY name")->fetchAll(\PDO::FETCH_COLUMN);
-            \ray($tables);
         } catch (\Exception $e) {
             \ray($e)->red();
+
             return Command::FAILURE;
         }
 
         return Command::SUCCESS;
+    }
+
+    protected function resetDatabase(): void
+    {
+        if (file_exists(SQLITE_FILE_PATH)) {
+            \unlink(SQLITE_FILE_PATH);
+        }
+
+        \touch(SQLITE_FILE_PATH);
     }
 }
